requirements.txt
# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
python-dateutil==2.8.2

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Additional modern dependencies
httpx==0.25.2
aiofiles==23.2.1 